修改VO层价格类型——221801420郑耀烽
修改VO层价格类型——221801420郑耀烽

+++ fidledemo/src/main/java/com/example/fidledemo/VO/GoodsItemVO.java
@@ -4,7 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -23,9 +23,9 @@ public class GoodsItemVO {
 
     private String title;
 
-    private Double price;
+    private BigDecimal price;
 
-    private Double originalPrice;
+    private BigDecimal originalPrice;
 
     private String imageLink;
