+++ fidledemo/src/main/java/com/example/fidledemo/VO/TaskVO.java
@@ -4,7 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -20,7 +20,7 @@ public class TaskVO {
 
     private Long pubId;
 
-    private Double reward;
+    private BigDecimal reward;
 
     private String title;
